// lib/enums.ts
export type AddressType = "" | "EX" | "SMTP" | "FAX" | "MHS" | "PROFS" | "X400"

export const enum MapiObjectType {
    // Address book container object
    MAPI_ABCONT = 4,
    // Address book object
    MAPI_ADDRBOOK = 2,
    // Message attachment object
    MAPI_ATTACH = 7,
    // Distribution list object
    MAPI_DISTLIST = 8,
    // Folder object
    MAPI_FOLDER = 3,
    // Form object
    MAPI_FORMINFO = 12,
    // Messaging user object
    MAPI_MAILUSER = 6,
    // Message object
    MAPI_MESSAGE = 5,
    // Profile section object
    MAPI_PROFSECT = 9,
    // Session object
    MAPI_SESSION = 11,
    // Status object
    MAPI_STATUS = 10,
    // Message store object
    MAPI_STORE = 1,
}

export const enum RecipientRowDisplayType {
    // A messaging user
    MessagingUser = 0x00,
    // A distribution list
    DistributionList = 0x01,
    // A forum, such as a bulletin board service or a public or shared folder
    Forum = 0x02,
    // An automated agent
    AutomatedAgent = 0x03,
    // An Address Book object defined for a large group, such as helpdesk, accounting, coordinator, or
    // department
    AddressBook = 0x04,
    // A private, personally administered distribution list
    PrivateDistributionList = 0x05,
    // An Address Book object known to be from a foreign or remote messaging system
    RemoteAddressBook = 0x06,
}

export const enum RecipientType {
    // The recipient is the message originator
    Originator = 0x0000,
    // The recipient is a primary (To) recipient. Clients are required to handle primary recipients. All other types are optional.
    To = 0x0001,
    // The recipient is a carbon copy (CC) recipient, a recipient that receives a message in addition to the primary recipients.
    Cc = 0x0002,
    // The recipient is a blind carbon copy (BCC) recipient. Primary and carbon copy recipients are unaware of the existence of BCC recipients.
    Bcc = 0x0003,
    // The recipient is a resource (e.g. a room)
    Resource = 0x0004,
    // The recipient is a room (uses PR_RECIPIENT_TYPE_EXE) needs Exchange 2007 or higher
    Room = 0x0007,
}